Fixes #1, #2  + Folder cleanup + Removal of input texture
Issues #1 (Crash at FluidInputOutput) and #2 (Issue with item I/O) are fixed.
Made a separate folder for capability stuff.
Item I/O is now done via a separate handler, so using the correct face for the capability is no longer needed. Because of this I've also removed the input textures, as these are now obsolete.