Added Ancite Big Bricks and Preservation Liquid & Rock
Added Ancite Big Bricks, which is purely decorative, as well as Preservation Liquid, which gives Regeneration II, Health Boost III and Water breather I to any entity swimming in it, except evil mobs. Those get Wither II.
In addition to that, when in contact with Water, Preservation Liquid turns into Preservation Rock (currently decorative only). In the case of Lava, it gets explodey.